Moved and renamed user.py into helpers
